+++ vscode/e2e/utils/vscody/fixture/index.ts
@@ -30,8 +30,6 @@ export interface TestContext {
     mitmProxy: MitMProxy
     //sourcegraphMitM: { endpoint: string; target: string }
     workspaceDir: Directory
-    //TODO(rnauta): Make the typing inferred from VSCode directly
-    executeCommand: <T = any>(commandId: string, ...args: any[]) => Promise<T>
 }
 
 export const fixture = mergeTests(
