+++ vscode/e2e/utils/vscody/uix/vscode/index.ts
@@ -1,0 +1,1 @@
+export { Session } from './internal'
